requirements.txt
aiohttp==3.10.10
discord.py==2.4.0
toml==0.10.2
